add refactor paths and rutes in project
