--- src/app/pokemons/components/index.ts
@@ -1,6 +1,0 @@
-
-/*
- * Exports components 
- */
-
-export { PokemonGrid } from './pokemonGrid/PokemonGrid';
