+++ src/pokemons/components/index.ts
@@ -1,0 +1,7 @@
+
+/*
+ * Exports components 
+ */
+
+export { PokemonGrid } from './pokemonGrid/PokemonGrid';
+export { PokemonCard } from './pokemonCard/PokemonCard';
