+++ src/pokemons/components/pokemonCard/PokemonCard.tsx
