+++ src/pokemons/components/pokemonGrid/PokemonGrid.tsx
@@ -8,11 +8,11 @@ interface Props  {
 
 export const PokemonGrid = ({ pokemons }:Props) => {
   return (
-    <div className='flex flex-wrap items-center justify-center'>
+   <div className="flex flex-wrap gap-10 items-center justify-center">
         {
-            pokemons.map(pokemon => (
-                <PokemonCard key={pokemon.id} pokemon={pokemon}/>
-            ))
+          pokemons.map( pokemon => (
+            <PokemonCard key={ pokemon.id } pokemon={pokemon} />            
+          ))
         }
     </div>
   )
