+++ src/pokemons/interfaces/simple-pokemon.ts
