+++ src/shopping-cart/components/counter/Counter.tsx
